Move resource allocator and add comemnt explaining cypher logic
